Add '/dadjoke' command
This command checks the message before the command call for a potential
dad joke. If it finds it, it makes the joke and deletes the command call.
If not, BreadBot calls out the command caller for making a bad joke.

This uses case insensitive regex matching to find the "I'm" sequence at the
beginning of a message.