Fix logging for Message and Ready events
The "ready" event would never get hit because it would not be registered
by the manager before login was called, so reverse the order on the
registration to allow for the event to be hit when necessary.

Additionally, add some logging for the "message" event to have verbose
tracking of users who've called commands and how.