Adjust bot presence setting to use envvars
Client is the only object that's run after dotenv.config()
is called in our program, so ideally we need to set envvar settings
here. This is fine, since we can still define types for our settings,
but just leave them blank and define them when initializing the client
with our envvars.