Lint entire rewrite under Airbnb style guide
I guess the JS files were linted, but the new rewrite certainly wasn't.
Add linting to all TS files and adjust accordingly.

The most notable pain point in the linting was the replacement of
`require()` with `import()`, which was Promise-based and essentially
messed with the original code format a lot. Fixing it was not
impossible, however.