Port "cat" and "dog" commands
By far the most popular commands on BreadBot. These are probably the
second hardest, since they're not as bad on API calls as the "acmurl"
routes will be.

The "any" typecast for Got responses is quite literally unavoidable
since I don't have types for the API's I'm hitting, so this is easiest
to deal with.