Fix InteractionCreate event to avoid fetch errors
This removes the ability to see arguments,
but it's impossible to dump the entire interaction to disk without
doing extreme recursion to fetch
all cached objects within it.

Arguably, one could iterate, but
Discord.js doesn't help here.

For moderation purposes, this is fine, since at most
it'll be the Echo command that gets abused, and in those cases
the commands can be tracked down
manually, whenever they occur.